+++ rusl/src/process.rs
@@ -1,6 +1,7 @@
 pub use clone::{clone, clone3, fork};
 pub use execve::execve;
 pub use exit::exit;
+pub use get_pid::get_pid;
 pub use signal::{add_signal_action, CatchSignal, SaSignalaction, SigInfo};
 pub use wait::wait_pid;
 
@@ -8,6 +9,7 @@ mod execve;
 mod exit;
 
 mod clone;
+mod get_pid;
 mod signal;
 #[cfg(test)]
 mod test;