+++ tiny-std/src/allocator/dlmalloc.rs
@@ -1455,7 +1455,7 @@ impl Dlmalloc {
             released += self.release_unused_segments();
 
             if released == 0 && self.topsize > self.trim_check {
-                self.trim_check = usize::max_value();
+                self.trim_check = usize::MAX;
             }
         }