+++ tiny-cli/Changelog.md
@@ -11,6 +11,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+## [v0.3.0] - 2024-08-24
+
+### Fixed
+- Some aliases not registering properly
+
+### Added
+- Positional arguments now accepted, and the default for untagged fields
+
+### Changed
+- Bools are always optional and should never be specified as `True` or `False`
+- Options must now be specified as either `long` or `short` with their alias
+
 ## [v0.2.1] - 2024-05-05
 ### Fixed